Refactored static functions off SnapGridComponent to MapGrid. Refactored away usages of SnapGridComponent.Position.
