+++ Robust.Shared/Map/MapGrid.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.Contracts;
+using System.Linq;
 using Robust.Shared.GameObjects;
 using Robust.Shared.Maths;
 using Robust.Shared.Timing;
@@ -9,7 +11,7 @@ using Robust.Shared.ViewVariables;
 namespace Robust.Shared.Map
 {
     /// <inheritdoc />
-    internal class MapGrid : IMapGridInternal
+    public class MapGrid : IMapGridInternal
     {
         /// <summary>
         ///     Game tick that the map was created.
@@ -547,5 +549,111 @@ namespace Robust.Shared.Map
         }
 
         #endregion Transforms
+
+        /// <summary>
+        ///     Returns an enumerable over all the entities which are one tile over in a certain direction.
+        /// </summary>
+        public static IEnumerable<IEntity> GetInDir(SnapGridComponent snapComp, Direction dir)
+        {
+            var transform = snapComp.Owner.Transform;
+            if (!snapComp._mapManager.TryGetGrid(transform.GridID, out var grid))
+                return Enumerable.Empty<IEntity>();
+
+            var pos = SnapGridPosAt(grid.SnapGridCellFor(transform.Coordinates), dir);
+
+            return grid.GetSnapGridCell(pos).Select(s => s.Owner);
+        }
+
+        [Pure]
+        public static IEnumerable<IEntity> GetOffset(SnapGridComponent snapComp, Vector2i offset)
+        {
+            var transform = snapComp.Owner.Transform;
+            if (!snapComp._mapManager.TryGetGrid(transform.GridID, out var grid))
+                return Enumerable.Empty<IEntity>();
+
+            var pos = grid.SnapGridCellFor(transform.Coordinates) + offset;
+
+            return grid.GetSnapGridCell(pos).Select(s => s.Owner);
+        }
+
+        public static IEnumerable<IEntity> GetLocal(SnapGridComponent snapComp)
+        {
+            var transform = snapComp.Owner.Transform;
+            if (!snapComp._mapManager.TryGetGrid(transform.GridID, out var grid))
+                return Enumerable.Empty<IEntity>();
+
+            return grid.GetSnapGridCell(grid.SnapGridCellFor(transform.Coordinates)).Select(s => s.Owner);
+        }
+
+        public static EntityCoordinates DirectionToGrid(SnapGridComponent snapComp, Direction direction)
+        {
+            var transform = snapComp.Owner.Transform;
+            if (!snapComp._mapManager.TryGetGrid(transform.GridID, out var grid))
+                return transform.Coordinates.Offset(direction.ToVec());
+
+            return grid.GridTileToLocal(SnapGridPosAt(grid.SnapGridCellFor(transform.Coordinates), direction));
+        }
+
+        private static Vector2i SnapGridPosAt(Vector2i position, Direction dir, int dist = 1)
+        {
+            switch (dir)
+            {
+                case Direction.East:
+                    return position + new Vector2i(dist, 0);
+                case Direction.SouthEast:
+                    return position + new Vector2i(dist, -dist);
+                case Direction.South:
+                    return position + new Vector2i(0, -dist);
+                case Direction.SouthWest:
+                    return position + new Vector2i(-dist, -dist);
+                case Direction.West:
+                    return position + new Vector2i(-dist, 0);
+                case Direction.NorthWest:
+                    return position + new Vector2i(-dist, dist);
+                case Direction.North:
+                    return position + new Vector2i(0, dist);
+                case Direction.NorthEast:
+                    return position + new Vector2i(dist, dist);
+                default:
+                    throw new NotImplementedException();
+            }
+        }
+
+        public static IEnumerable<SnapGridComponent> GetCardinalNeighborCells(SnapGridComponent snapComp)
+        {
+            var transform = snapComp.Owner.Transform;
+            if (!snapComp._mapManager.TryGetGrid(transform.GridID, out var grid))
+                yield break;
+
+            var position = grid.SnapGridCellFor(transform.Coordinates);
+            foreach (var cell in grid.GetSnapGridCell(position))
+                yield return cell;
+            foreach (var cell in grid.GetSnapGridCell(position + new Vector2i(0, 1)))
+                yield return cell;
+            foreach (var cell in grid.GetSnapGridCell(position + new Vector2i(0, -1)))
+                yield return cell;
+            foreach (var cell in grid.GetSnapGridCell(position + new Vector2i(1, 0)))
+                yield return cell;
+            foreach (var cell in grid.GetSnapGridCell(position + new Vector2i(-1, 0)))
+                yield return cell;
+        }
+
+        public static IEnumerable<SnapGridComponent> GetCellsInSquareArea(SnapGridComponent snapComp, int n = 1)
+        {
+            var transform = snapComp.Owner.Transform;
+            if (!snapComp._mapManager.TryGetGrid(transform.GridID, out var grid))
+                yield break;
+
+            var position = grid.SnapGridCellFor(transform.Coordinates);
+
+            for (var y = -n; y <= n; ++y)
+            for (var x = -n; x <= n; ++x)
+            {
+                foreach (var cell in grid.GetSnapGridCell(position + new Vector2i(x, y)))
+                {
+                    yield return cell;
+                }
+            }
+        }
     }
 }
